+++ biopipen/ns/scrna.py
@@ -146,6 +146,9 @@ class SeuratClusterStats(Proc):
             nCellsPerSample - Number of cells per sample for each cluster
             percCellsPerSample - Percentage of cells per sample for each cluster
         exprs: The expression values to plot
+            genes - The set of genes for the plots, unless `features` for those
+            plots is specified. Could also specify a file with genes
+            (one per line)
             ridgeplots - The ridge plots for the gene expressions.
             See `?Seurat::RidgePlot`.
             vlnplots - Violin plots for the gene expressions.
@@ -217,32 +220,17 @@ class CellsDistribution(Proc):
             >>> # If not given, will use `{{in.srtobj | stem}}`, ...
             >>> name = ""
             >>> [cases.case1]
-            >>> #
-            >>> [cases.case1.group]
-            >>> # Use values of a column from metadata for the groups
-            >>> by = "Timepoint"
-            >>> # or use mutaters to add a column to metadata
-            >>> # Specify the order of the groups to show on the plot
-            >>> # from left to right (columns)
-            >>> # Optional
-            >>> order = ["CONTROL", "Responder", "Non-Responder"]
-            >>> [cases.case1.group.mutaters]
+            >>> filter = 'Responder != "Other"'
+            >>> # Create some helper columns
+            >>> [cases.case1.mutaters]
             >>> Responder = '''
             >>>     case_when(
             >>>         PFS == "CONTROL" ~ "CONTROL",
             >>>         PFS >= 60 ~ "Responder",
-            >>>         TRUE ~ "Non-Responder"
+            >>>         PFS < 60 ~ "Non-Responder",
+            >>>         TRUE ~ "Other"
             >>>     )
             >>> '''
-            >>> #
-            >>> [cases.case1.cells]
-            >>> by = "CDR3.aa" # "CloneSize"
-            >>> # Specify the order of the cell identities to show on the plot
-            >>> # The order of the cells (rows)
-            >>> orderby = "desc(.CloneSize)"
-            >>> n = 10
-            >>> # or use mutaters to add a column to metadata
-            >>> [cases.case1.cells.mutaters]
             >>> CloneSize = '''
             >>>     case_when(
             >>>         Clones > 50 ~ "Clone50p",
@@ -251,6 +239,20 @@ class CellsDistribution(Proc):
             >>>         TRUE ~ "CloneSmall"
             >>>     )
             >>> '''
+            >>> [cases.case1.group]
+            >>> # Use values of a column from metadata for the groups
+            >>> by = "Timepoint"
+            >>> # Specify the order of the groups to show on the plot
+            >>> # from left to right (columns)
+            >>> # Optional
+            >>> order = ["CONTROL", "Responder", "Non-Responder"]
+            >>> #
+            >>> [cases.case1.cells]
+            >>> by = "CDR3.aa" # "CloneSize"
+            >>> # Specify the order of the cell identities to show on the plot
+            >>> # The order of the cells (rows)
+            >>> orderby = "desc(.CloneSize)"
+            >>> n = 10
 
     Output:
         outdir: The output directory
@@ -265,6 +267,12 @@ class CellsDistribution(Proc):
         - name: r-seurat
           check: |
               {{proc.lang}} -e "library(Seurat)"
+        - name: r-dplyr
+          check: |
+              {{proc.lang}} -e "library(dplyr)"
+        - name: r-tidyr
+          check: |
+              {{proc.lang}} -e "library(tidyr)"
     """
     input = "srtobj:file, casefile:file"
     output = "outdir:dir:{{in.srtobj | stem}}.cells_distribution"
@@ -409,6 +417,8 @@ class MarkersFinder(Proc):
             >>> "ident.1" = "Tumor"
             >>> "ident.2" = "Normal"
             >>> "group.by" = "Source"
+            >>> # focus on a subset of cells
+            >>> filter = "SampleType != 'Control'"
             >>> # other arguments for Seruat::FindMarkers()
 
             We can also use a new group.by:
@@ -417,6 +427,8 @@ class MarkersFinder(Proc):
             >>> "ident.2" = "Control"
             >>> "group.by" = "Group"
             >>> # other arguments for Seruat::FindMarkers()
+            >>> # Filter after mutaters
+            >>> filter2 = "SampleType != 'Control'"
             >>> [cases.case2.mutaters]
             >>> Group = '''
             >>>   if_else(Source %in% c("Tumor", "Normal"), "Case", "Control")
@@ -676,6 +688,7 @@ class ScFGSEA(Proc):
             See `in.casefile` from `scrna.MarkersFinder`
             `ident.2` is required for each case.
             One could also use placeholders for the cases.
+            To enable this, you need `percluster = True` in the config.
             Currently only cluster is supported. One could use `{cluster}` or
             `{ident}` to denote the clusters.