+++ biopipen/scripts/scrna/MarkersFinder.R
@@ -7,6 +7,7 @@ library(tibble)
 library(Seurat)
 library(enrichR)
 library(future)
+library(tidyseurat)
 
 setEnrichrSite("Enrichr")
 
@@ -83,8 +84,15 @@ do_case = function(case) {
     print(paste("- Dealing with case:", case, "..."))
     casepms = cases$cases[[case]]
     pmnames = names(casepms)
-    obj = mutate_meta(seurat_obj, casepms$mutaters)
+    obj = seurat_obj
+    if ("filter" %in% pmnames) {
+        obj = obj |> filter(eval(parse(text=casepms$filter)))
+    }
+    obj = mutate_meta(obj, casepms$mutaters)
     casepms$mutaters = NULL
+    if ("filter2" %in% pmnames) {
+        obj = obj |> filter(eval(parse(text=casepms$filter2)))
+    }
 
     if (!"ident.1" %in% pmnames && !"ident.2" %in% pmnames) {
         Idents(obj) = casepms$group.by